Start tui implementation
This adds basic functionality to list nodes and profiles, showing the
state of nodes and starting a node (with full CLI configuration options).

The TUI is not yet at parity with the CLI and will likely require a bit
of work to keep in sync, but it look promising.

One challenge remaining is that of gracefully handling viewport
overflow. While this could maybe be done via ctrl+arrow keys (rendering
to a larger, virtual screen and then displaying a subset), it would not
be ideal as some sense of visible elements might be necessary and
likewise, focused elements should be visible (and thus, scrolled
automatically).

This will require some additional thought, but for now, this is a
starting point.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>