Add Transit-based auto-unseal to startup
This adds a special flag, --seals, to the CLI in order to allow
auto-unseal configuration to be added to persistent nodes. This uses a
URI scheme (currently only supporting Transit) to identify the target of
auto-unseal:

> --seals http(s)://<TOKEN>@<HOST>/<MOUNT>/keys/<UNSEAL-KEY>

where TOKEN is the desired OpenBao token, <HOST> is the address of the
instance including any port number, MOUNT is the transit mount in
question, and UNSEAL-KEY is the name of the key to be used for seal and
unseal.

This results in using transit auto-unseal, as demonstrated by:

$ devbao node stop prod
$ devbao node resume prod
$ bao secrets list ## will not report the instance is sealed

However, it can still be manually sealed and subsequently unsealed using
the recovery keys like a regular node:

$ devbao node seal
$ devbao node unseal

(or with the new --unseal flag to ignore the behavior difference between
 auto-unseal and shamir's sealed nodes).

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>