Add userpass profile
This profile adds a userpass authentication method, useful for managing
clustered instances of Vault Enterprise (as a root token on one cluster
may not work on another, but mounts are replicated). For OpenBao's HA
mode, this matters less, but does help with simulating actual user
behavior.

We add two user accounts:

 - u: testing/p: testing, simulating a restricted test user with access
   to PKI & Transit, and
 - u: admin/p: admin, for getting the equivalent of a root token via
   password login (using a very broadly permissive policy).

Ideally the scope of the former would increase as more profiles are
added.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>