Add cluster build command
This adds the initial code for building clusters, letting callers take a
single, isolated node and start building a cluster around it. Currently
the only (implicitly) supported type of cluster will be a Seal HA
cluster.

This also updates node listing to show which cluster a node is in, if
any.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>