Add cluster join command
This allows adding a node to a HA cluster, automatically updating its
state (root token, unseal or recovery keys) and unsealing the node. This
is only performed if the two node's seal configurations can be proven to
be equivalent (either via Transit auto-unseal or implicit Shamir's).
Leadership should not be affected by this operation.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>