Add cluster remove command
This allows removing nodes from a HA cluster automatically, updating the
cluster's state. When building removing nodes, their state is still
persisted in their local storage; on restart, they will continue
proxying requests to the primary node of the cluster if is up, but the
cluster will not see it (in either sys/ha-status after restarting the
cluster's nodes or in sys/storage/raft/configuration) and thus the node
will not be able to vote in leadership elections.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>